* ICartService ( type Entity change to ContentClassForUnitTest) * type of identifier Id changed to ulong * AddNewContentInCartUnitTests (change logic methods Add_New_Content_In_Cart and and Add_New_Content_In_Cart_If_Not_Save_In_Repository + change int to ulong identifier) * Models ContentCart, ContentCartDto (delete property from the abstract class and improved properties for business logic) * Change name file AddNewContentInCartUnitTests + UserClassForUnitTest for the get userId

+++ MediaShop.Common/Interfaces/Repositories/ICartRespository.cs
@@ -19,7 +19,7 @@
         /// </summary>
         /// <param name="id">identificator</param>
         /// <returns>rezalt operation</returns>
-        TModel Get(int id);
+        TModel Get(ulong id);
 
         /// <summary>
         /// Method for updating object type ContentCart
@@ -40,7 +40,7 @@
         /// </summary>
         /// <param name="id">identificator</param>
         /// <returns>rezalt operation</returns>
-        TModel Delete(int id);
+        TModel Delete(ulong id);
 
         /// <summary>
         /// Method for find collection of object type ContentCart
