+++ MediaShop.Common/Models/Entity.cs
@@ -10,7 +10,7 @@
         /// <summary>
         /// Gets or sets object identificator
         /// </summary>
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         /// <summary>
         /// Gets or sets date of object creation
@@ -20,7 +20,7 @@
         /// <summary>
         /// Gets or sets cart creator Id
         /// </summary>
-        public int CreatorId { get; set; }
+        public ulong CreatorId { get; set; }
 
         /// <summary>
         /// Gets or sets date of object modification
@@ -30,6 +30,6 @@
         /// <summary>
         /// Gets or sets object modificator
         /// </summary>
-        public int? ModifierId { get; set; }
+        public ulong? ModifierId { get; set; }
     }
 }
